// tools/gitignore-generator/client/types.ts
export type GitignoreGeneratorMessages = Readonly<{
  meta: {
    name: string
    description: string
  }
  templatesLabel: string
  templatesDescription: string
  searchLabel: string
  searchDescription: string
  searchPlaceholder: string
  quickSelectLabel: string
  selectedTemplatesLabel: string
  selectedCountLabel: string
  clearSelectionLabel: string
  emptySelectionLabel: string
  noTemplatesFoundLabel: string
  languagesLabel: string
  globalLabel: string
  communityLabel: string
  resultLabel: string
  resultDescription: string
  previewPlaceholder: string
  copyResultLabel: string
  copiedLabel: string
  downloadGitignoreLabel: string
}>